Move Header, Sidebar, Footer and PageTemplate components into their own folder, and update frontend page imports
