+++ pages/images/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { PageTemplate } from 'components/PageTemplate'
+import { PageTemplate } from 'components/main/PageTemplate'
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSearch } from "@fortawesome/free-solid-svg-icons";
 import { Col, Row, Form, InputGroup, Button} from '@themesberg/react-bootstrap';
